Conflicts:
	pom.xml
	src/main/java/org/atlasapi/remotesite/pa/PaModule.java
	src/main/java/org/atlasapi/remotesite/pa/data/DefaultPaProgrammeDataStore.java
	src/main/java/org/atlasapi/remotesite/pa/data/PaProgrammeDataStore.java

+++ src/main/webapp/static/js/applications.js
@@ -263,9 +263,17 @@ var updatePrecedence = function(callback) {
 var updateEnabled = function(callback){
 	var slug = app.slug;
 	var count = 0;
+	
+	
 	for(var i = 0, ii = app.configuration.publishers.length; i<ii; i++){
 		var enabled = app.configuration.publishers[i].enabled;
 		var publisher = app.configuration.publishers[i].key;
+		var available = app.configuration.publishers[i].state == "available";
+
+		if (!available) {
+			count++;
+			continue;
+		}
 		var type = 'delete';
 		var data = null;
 		if(enabled){
@@ -282,6 +290,7 @@ var updateEnabled = function(callback){
 			type: type,
 			url: url,
 			data: data,
+			async: false,
 			success:function(responseData, textStatus, XMLHttpRequest) {
 				
 			},
@@ -296,6 +305,8 @@ var updateEnabled = function(callback){
 			}
 		});
 	}
+	
+	
 }
 
 var disablePrecedence = function(callback) {
@@ -378,4 +389,19 @@ var page = {
 		
 		$('#app-publishers tbody').replaceWith(publisherString);
 	}
-};	
+};
+
+var runSearch = function() {
+	var url = window.location.pathname + "?search=" + encodeURIComponent($("#search").val());	
+	window.location.href = url;
+};
+
+$("#runSearch").live('click', function() {
+	runSearch();
+});
+
+$('#search').live('keypress',function(e) {
+	if (e.which == 13) {
+		runSearch();	
+	}
+});
